refactor(CardDeck): CardDeck 추상화 작업
