+++ src/main/kotlin/blackjack/domain/carddeck/CardDeck.kt
@@ -1,0 +1,8 @@
+package blackjack.domain.carddeck
+
+import blackjack.domain.Card
+
+interface CardDeck {
+
+    fun drawCard(): Card
+}
