+++ src/main/kotlin/blackjack/domain/carddeck/RandomCardDeck.kt
@@ -1,12 +1,14 @@
-package blackjack.domain
+package blackjack.domain.carddeck
 
 import blackjack.Shape
+import blackjack.domain.Card
+import blackjack.domain.CardNumber
 
-object CardDeck {
+object RandomCardDeck : CardDeck {
 
     private val Cards: List<Card> = Shape.values().fold(listOf()) { cards, shape -> cards + makeCard(shape) }
 
     private fun makeCard(shape: Shape): List<Card> = CardNumber.values().map { Card(shape, it) }
 
-    fun drawCard(): Card = Cards.shuffled().first()
+    override fun drawCard(): Card = Cards.shuffled().first()
 }
