+++ src/test/kotlin/blackjack/domain/ParticipantsTest.kt
@@ -1,0 +1,3 @@
+package blackjack.domain
+
+class ParticipantsTest
